feat: move test order status in integration test
